feat: enhance portfolio section with detailed architecture, performance metrics, tech stacks, and workflow diagrams
