+++ components/aboutPortoPage/index.ts
@@ -4,5 +4,8 @@
  * @copyright 2025 ColdByDefault. All Rights Reserved.
  */
 
-export * from "./portfolio-section";
-
+export { ArchitectureDiagram } from "./portfolio-architecture";
+export { PerformanceMetrics } from "./portfolio-performance";
+export { TechStackGrid } from "./portfolio-tech";
+export { WorkflowDiagram } from "./portfolio-workflow";
+export { CodeExamples } from "./portfolio-showcase";
